reduce float casting inspect outputs

+++ pmsp/dataset.py
@@ -7,9 +7,9 @@ from torch.utils.data import Dataset
 
 class PMSPDataset(Dataset):
     def __init__(self, df):
-        self.frequency_tensor = torch.tensor(df["frequency"])
-        self.grapheme_tensor = torch.tensor(df["graphemes"])
-        self.phoneme_tensor = torch.tensor(df["phonemes"])
+        self.frequency_tensor = torch.tensor(df["frequency"]).float()
+        self.grapheme_tensor = torch.tensor(df["graphemes"]).float()
+        self.phoneme_tensor = torch.tensor(df["phonemes"]).float()
         self.num_samples = len(df)
 
     def __len__(self):
